+++ src/vocabulary/colorNames.js
@@ -1,0 +1,27 @@
+export const colorNames = [
+    { word: "赤", answer: "あか", meaning: "le rouge" },
+    { word: "青", answer: "あお", meaning: "le bleu" },
+    { word: "黄色", answer: ["きいろ", "オウショク"], meaning: "le jaune" },
+    { word: "l'orange", answer: ["オレンジ色", "橙色", "だいだいいろ"] },
+    { word: "紫", answer: "むらさき", meaning: "le violet" },
+    { word: "緑", answer: "みどり", meaning: "le vert" },
+    { word: "白", answer: "しろ", meaning: "le blanc" },
+    { word: "黒", answer: "くろ", meaning: "le noir" },
+    { word: "灰色", answer: ["はいいろ", "カイショク"], meaning: "le gris" },
+    { word: "le rose", answer: "ピンク" },
+    { word: "茶色", answer: "チャいろ", meaning: "le marron" },
+    { word: "le beige", answer: "ベージュ" },
+    { word: "薄紫色", answer: ["うすむらさきいろ", "モーブ"], meaning: "le mauve" },
+    { word: "le magenta", answer: "マゼンタ" },
+    { word: "紅", answer: "くれない", meaning: "le cramoisi" },
+    { word: "le cyan", answer: "シアン" },
+    { word: "水色", answer: "みずいろ", meaning: "le bleu clair" },
+    { word: "l'indigo", answer: "インディゴ" },
+    { word: "虹色", answer: "にじいろ", meaning: "couleur arc-en-ciel" },
+    { word: "金色", answer: "キンいろ", meaning: "l'or" },
+];
+
+export default {
+    list: colorNames,
+    ref: "https://www.youtube.com/watch?v=MvaEDg_Hwusat=0s",
+};