+++ src/vocabulary/familyMembers.js
@@ -1,0 +1,47 @@
+export const familyMembers = [
+    { word: "夫", answer: "おっと", meaning: "mon époux" },
+    { word: "妻", answer: "つま", meaning: "mon épouse" },
+    { word: "両親", answer: "リョウシン", meaning: "mes parents" },
+    { word: "父", answer: "ちち", meaning: "mon père" },
+    { word: "母", answer: "はは", meaning: "ma mère" },
+    { word: "子供", answer: "こども", meaning: "mon enfant" },
+    { word: "息子", answer: "むすこ", meaning: "mon fils" },
+    { word: "娘", answer: "むすめ", meaning: "ma fille" },
+    { word: "祖父母", answer: "ソフボ", meaning: "mes grands-parents" },
+    { word: "祖父", answer: "ソフ", meaning: "mon grand-père" },
+    { word: "祖母", answer: "ソボ", meaning: "ma grand-mère" },
+    { word: "孫", answer: "まご", meaning: "mon petit-enfant" },
+    { word: "双子", answer: "ふたご", meaning: "mon jumeau / ma jumelle" },
+
+    { word: "兄", answer: "あに", meaning: "mon frère aîné" },
+    { word: "弟", answer: "おとうと", meaning: "mon frère cadet" },
+    { word: "兄弟", answer: "キョウダイ", meaning: "la fratrie" },
+
+    { word: "姉", answer: "あね", meaning: "ma sœur aînée" },
+    { word: "妹", answer: "いもうと", meaning: "ma sœur cadette" },
+
+    { word: "mon cousin", answer: "いとこ" },
+    { word: "mon cousin (aîné)", answer: "従兄", meaning: "いとこ" },
+    { word: "mon cousin (cadet)", answer: "従弟", meaning: "いとこ" },
+    { word: "ma cousine (aîné)", answer: "従姉", meaning: "いとこ" },
+    { word: "ma cousine (cadette)", answer: "従妹", meaning: "いとこ" },
+
+    { word: "mon oncle", answer: "おじ" },
+    { word: "mon oncle (aîné)", answer: "伯父", meaning: "おじ" },
+    { word: "mon oncle (cadet)", answer: "叔父", meaning: "おじ" },
+
+    { word: "ma tante", answer: "おば" },
+    { word: "ma tante (aînée)", answer: "伯母", meaning: "おば" },
+    { word: "ma tante (cadette)", answer: "叔母", meaning: "おば" },
+
+    { word: "義兄", answer: "ギケイ", meaning: "mon beau-frère (aîné)" },
+    { word: "義弟", answer: "ギテイ", meaning: "mon beau-frère (cadet)" },
+
+    { word: "義姉", answer: "ギシ", meaning: "ma belle-sœur (aînée)" },
+    { word: "義妹", answer: "ギマイ", meaning: "ma belle-sœur (cadette)" },
+];
+
+export default {
+    list: familyMembers,
+    ref: "https://www.youtube.com/watch?v=KJak7DtVtwI&t=0s"
+}