+++ src/vocabulary/index.js
@@ -1,6 +1,8 @@
+import colorNames from "./colorNames";
 import countryNames from "./countryNames";
 import domesticAnimals from "./domesticAnimals";
 import domesticEquipment from "./domesticEquipment";
+import familyMembers from "./familyMembers";
 import firstKanji from "./firstKanji";
 import numbers from "./numbers";
 import politeness from "./politeness";
@@ -11,5 +13,7 @@ export default {
     "Le Matériel Domestique": domesticEquipment,
     "Le Nom des Pays": countryNames,
     "Les Animaux Domestiques": domesticAnimals,
+    "Les Membres de la Famille": familyMembers,
+    "Le Nom des Couleurs": colorNames,
     "Les Nombres": numbers,
 };